fix: fix current user error
