+++ src/components/layouts/TweetLayout.jsx
@@ -1,13 +1,14 @@
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 import { Outlet, useLocation } from 'react-router-dom';
 import { GridContainer } from '../../globalStyles';
 import { Navbar } from '../Navbar';
 import Popular from '../Popular';
 import { useUser } from '../../contexts/UserContext';
 import { addTweet } from '../../api/tweet';
+import { getCurrentUser, getUserFollowings } from '../../api/user';
 
 export default function TweetLayout() {
-  const { currentUser } = useUser();
+  const { currentUser, setCurrentUser, setUserFollowings } = useUser();
   const [tweets, setTweets] = useState([]);
   const [tweetInput, setTweetInput] = useState('');
   const { pathname } = useLocation();
@@ -51,6 +52,27 @@ export default function TweetLayout() {
       return { status: 'error' };
     }
   };
+
+  useEffect(() => {
+    const getUserAsync = async () => {
+      try {
+        const user = await getCurrentUser();
+        console.log(`user ${user.id} just logged in`);
+        const followings = await getUserFollowings(user.id);
+        const followingUsers = followings.map(
+          (following) => following.followingId
+        );
+        // followingUsers = [34, 44]
+        console.log('user following list get');
+        setCurrentUser(user);
+        setUserFollowings(followingUsers);
+        console.log('user context loaded');
+      } catch {
+        console.error(error);
+      }
+    };
+    getUserAsync();
+  }, []);
 
   return (
     <GridContainer>
